+++ R/getmzrt.R
@@ -634,6 +634,9 @@ getoverlaprt <- function(rtrange1, rtrange2) {
 #' @param n the number of equally spaced points at which the density is to be estimated, default 512
 #' @param log log transformation
 #' @return Density weighted intensity for one sample
+#' @examples
+#' data(list)
+#' getdwtus(list$data[,1])
 #' @export
 #'
 getdwtus <- function(peak,n=512,log=F){
@@ -642,4 +645,23 @@ getdwtus <- function(peak,n=512,log=F){
         }
         sum <- sum(density(peak,bw='sj',n=n)$x*density(peak,bw='sj',n=n)$y)
         return(sum)
+}
+
+#' Compute pooled QC linear index according to run order
+#' @param data peaks intensity list with row as peaks and column as samples
+#' @param order run order of pooled QC samples
+#' @param n samples numbers used for linear regression
+#' @return vector for the peaks proportion with significant changes in linear regression after FDR control.
+#' @export
+
+getpqsi <- function(data, order, n=5){
+        data <- data[,order(order)]
+        porp <- 1:ncol(data)
+        for(i in n:ncol(data)){
+                p <- apply(data[,c((i-n+1):i)],1,function(x) summary(stats::lm(x~c((i-n+1):i)))$coefficients[2,4])
+                # FDR control
+                q <- stats::p.adjust(p,method='BH')
+                porp[i] <- sum(q<0.1)/nrow(data)
+        }
+        return(porp[-c(1:(n-1))])
 }